complete refactoring
